Added carousel.reloadData and viewWillTransition to ManageSubscriptionVC
