+++ Story Squad/Story Squad/View Controllers/FamilySettingsViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import ScalingCarousel
 import CoreData
 
 class FamilySettingsViewController: UIViewController {
@@ -18,8 +17,6 @@ class FamilySettingsViewController: UIViewController {
     
     
     // MARK: - Outlets
-    @IBOutlet weak var carousel: ScalingCarouselView!
-    
     @IBOutlet weak var enterNewEmailTextField: UITextField!
     @IBOutlet weak var enterOldPasswordTextField: UITextField!
     @IBOutlet weak var enterNewPasswordTextField: UITextField!
@@ -30,22 +27,13 @@ class FamilySettingsViewController: UIViewController {
         
         receiveDataFromSignup()
     }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        carousel.reloadData()
-    }
-    
+
     // To receive the Parent and NetworkingController from the Tab Bar
     func receiveDataFromSignup() {
         guard let tabBar = tabBarController as? MainTabBarController else { return }
         
         self.parentUser = tabBar.parentUser
         self.networkingController = tabBar.networkingController
-    }
-
-    override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
-        super.viewWillTransition(to: size, with: coordinator)
-        carousel.deviceRotated()
     }
     
     @IBAction func addChildButtonTapped(_ sender: UIButton) {
@@ -64,38 +52,4 @@ class FamilySettingsViewController: UIViewController {
     }
 }
 
-
-extension FamilySettingsViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout, NSFetchedResultsControllerDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "childCollectionViewCell", for: indexPath)
-
-        DispatchQueue.main.async {
-            cell.setNeedsLayout()
-            cell.layoutIfNeeded()
-        }
-        
-        return cell
-    }
-    
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        carousel.didScroll()
-        
-       // guard let currentCenterIndex = carousel.currentCenterCellIndex?.row else { return }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        
-        return 10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        
-        return 0
-    }
-}
